fix(tests): merge API user tests into common API tests
